Move delete button to update page (#27)
* add delete workspace modal


* implement delete on workspace overview page


* fix export on delete workspace modal


* add try catch to handle errors for workspace delete


* move visibility control to workspace overview page exlusively


* remove unused import


* change workspace overview route to workspace update


* move delete button from workspace overview page to update page


* remove update button from workspace overview page


* recover router to workspace overview page


* change navigation url for workspace overview button on left side panel


---------
