+++ src/plugins/workspace/public/components/workspace_updater/workspace_updater.tsx
@@ -4,7 +4,14 @@
  */
 
 import React, { useCallback, useEffect, useState } from 'react';
-import { EuiPage, EuiPageBody, EuiPageHeader, EuiPageContent } from '@elastic/eui';
+import {
+  EuiPage,
+  EuiPageBody,
+  EuiPageHeader,
+  EuiPageContent,
+  EuiButton,
+  EuiPanel,
+} from '@elastic/eui';
 import { useObservable } from 'react-use';
 import { i18n } from '@osd/i18n';
 import { of } from 'rxjs';
@@ -20,6 +27,7 @@ import {
   WORKSPACE_OP_TYPE_UPDATE,
 } from '../../../common/constants';
 import { ApplicationStart } from '../../../../../core/public';
+import { DeleteWorkspaceModal } from '../delete_workspace_modal';
 
 export const WorkspaceUpdater = () => {
   const {
@@ -34,6 +42,7 @@ export const WorkspaceUpdater = () => {
   const { [excludedAttribute]: removedProperty, ...otherAttributes } =
     currentWorkspace || ({} as WorkspaceAttribute);
 
+  const [deleteWorkspaceModalVisible, setDeleteWorkspaceModalVisible] = useState(false);
   const [currentWorkspaceFormData, setCurrentWorkspaceFormData] = useState<
     Omit<WorkspaceAttribute, 'id'>
   >(otherAttributes);
@@ -71,7 +80,7 @@ export const WorkspaceUpdater = () => {
             defaultMessage: 'Update workspace successfully',
           }),
         });
-        application.navigateToApp(WORKSPACE_APP_ID, {
+        await application.navigateToApp(WORKSPACE_APP_ID, {
           path: PATHS.overview + '?' + WORKSPACE_ID_IN_SESSION_STORAGE + '=' + currentWorkspace.id,
         });
         return;
@@ -89,11 +98,51 @@ export const WorkspaceUpdater = () => {
   if (!currentWorkspaceFormData.name) {
     return null;
   }
+  const deleteWorkspace = async () => {
+    if (currentWorkspace?.id) {
+      let result;
+      try {
+        result = await workspaces?.client.delete(currentWorkspace?.id);
+      } catch (error) {
+        notifications?.toasts.addDanger({
+          title: i18n.translate('workspace.delete.failed', {
+            defaultMessage: 'Failed to delete workspace',
+          }),
+          text: error instanceof Error ? error.message : JSON.stringify(error),
+        });
+        return setDeleteWorkspaceModalVisible(false);
+      }
+      if (result?.success) {
+        notifications?.toasts.addSuccess({
+          title: i18n.translate('workspace.delete.success', {
+            defaultMessage: 'Delete workspace successfully',
+          }),
+        });
+      } else {
+        notifications?.toasts.addDanger({
+          title: i18n.translate('workspace.delete.failed', {
+            defaultMessage: 'Failed to delete workspace',
+          }),
+          text: result?.error,
+        });
+      }
+    }
+    setDeleteWorkspaceModalVisible(false);
+    await application.navigateToApp('home');
+  };
 
   return (
     <EuiPage paddingSize="none">
       <EuiPageBody panelled>
-        <EuiPageHeader restrictWidth pageTitle="Update Workspace" />
+        <EuiPageHeader
+          restrictWidth
+          pageTitle="Update Workspace"
+          rightSideItems={[
+            <EuiButton color="danger" onClick={() => setDeleteWorkspaceModalVisible(true)}>
+              Delete
+            </EuiButton>,
+          ]}
+        />
         <EuiPageContent
           verticalPosition="center"
           horizontalPosition="center"
@@ -102,6 +151,15 @@ export const WorkspaceUpdater = () => {
           hasShadow={false}
           style={{ width: '100%', maxWidth: 1000 }}
         >
+          {deleteWorkspaceModalVisible && (
+            <EuiPanel>
+              <DeleteWorkspaceModal
+                onConfirm={deleteWorkspace}
+                onClose={() => setDeleteWorkspaceModalVisible(false)}
+                selectedItems={currentWorkspace?.name ? [currentWorkspace.name] : []}
+              />
+            </EuiPanel>
+          )}
           {application && (
             <WorkspaceForm
               application={application}
