+++ src/plugins/workspace/public/plugin.ts
@@ -116,7 +116,7 @@ export class WorkspacesPlugin implements Plugin<{}, {}> {
     core.chrome.setCustomNavLink({
       title: i18n.translate('workspace.nav.title', { defaultMessage: 'Workspace Overview' }),
       baseUrl: core.http.basePath.get(),
-      href: core.application.getUrlForApp(WORKSPACE_APP_ID, { path: PATHS.overview }),
+      href: core.application.getUrlForApp(WORKSPACE_APP_ID, { path: PATHS.update }),
     });
     this._changeSavedObjectCurrentWorkspace();
     return {};
